refactor(fiber): split into Fiber and Fork
