--- src/internals/Tag.ts
@@ -1,11 +1,0 @@
-export enum Tag {
-  Constant = 0,
-  Resume = 3,
-  ResumeM = 4,
-  Map = 5,
-  Chain = 6,
-  Async = 7,
-  Reject = 8,
-  Never = 9,
-  Catch = 10
-}
