+++ src/runtimes/BaseRuntime.ts
@@ -5,8 +5,8 @@ import {ICancellable, IScheduler} from 'ts-scheduler'
 
 import {CancellationList} from '../internals/CancellationList'
 import {noop} from '../internals/Noop'
-import {Fiber} from '../main/Fiber'
 import {FIO} from '../main/FIO'
+import {Fork} from '../main/Fork'
 
 import {IRuntime} from './IRuntime'
 
@@ -23,7 +23,7 @@ export abstract class BaseRuntime<R> implements IRuntime<R> {
     const cancellationList = new CancellationList()
     cancellationList.push(
       this.scheduler.asap(
-        Fiber,
+        Fork,
         io.toFiber(),
         this.env,
         rej,
